requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
pymysql
boto3
python-dotenv
passlib[bcrypt]
google-generativeai
pdfplumber
reportlab
